# swagger.yaml
openapi: 3.0.0
info:
  title: URL Shortener API
  description: API para encurtamento de URLs, com autenticação de usuários e gerenciamento de URLs encurtadas.
  version: 1.0.0
  contact:
    name: Suporte
    email: suporte@dominio.com
    url: https://dominio.com

servers:
  - url: http://localhost:4130
    description: Servidor local para desenvolvimento

paths:
  /auth/login:
    post:
      summary: Realiza login do usuário
      description: Verifica as credenciais do usuário e retorna um token JWT.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: Email do usuário.
                  example: john@example.com
                password:
                  type: string
                  description: Senha do usuário.
                  example: "senha123"
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT para autenticação.
        '400':
          description: Credenciais inválidas
        '404':
          description: Usuário não encontrado
        '500':
          description: Erro interno do servidor

  /auth/register:
    post:
      summary: Registra um novo usuário
      description: Cria uma nova conta de usuário e retorna um token JWT.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nome do usuário.
                  example: John Doe
                email:
                  type: string
                  format: email
                  description: Email do usuário.
                  example: john@example.com
                password:
                  type: string
                  description: Senha do usuário.
                  example: "senha123"
      responses:
        '201':
          description: Registro bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT para autenticação.
        '400':
          description: Dados inválidos
        '409':
          description: Email já em uso
        '500':
          description: Erro interno do servidor

  /auth/refresh:
    post:
      summary: Gera um novo access token usando o refresh token
      description: Verifica o refresh token e, se válido, gera um novo access token.
      operationId: refreshAccessToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: O refresh token do usuário.
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Refresh token válido, novo access token gerado
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: Novo access token gerado.
        '401':
          description: Refresh token inválido ou expirado
        '500':
          description: Erro ao gerar novo access token

  /urls/shorten:
    post:
      summary: Encurta uma URL
      description: Este endpoint permite que qualquer pessoa encurte uma URL. Usuários autenticados terão a URL associada ao seu ID.
      operationId: shortenUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                  description: A URL que será encurtada.
      responses:
        '201':
          description: URL encurtada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  shortUrl:
                    type: string
                    description: A URL encurtada gerada.
        '400':
          description: URL inválida
        '500':
          description: Erro ao encurtar a URL

  /urls/redirect/{shortUrl}:
    get:
      summary: Redireciona para a URL original
      description: Redireciona para a URL original associada ao URL encurtado, além de contar o número de cliques.
      operationId: redirectUrl
      parameters:
        - name: shortUrl
          in: path
          required: true
          description: O código da URL encurtada.
          schema:
            type: string
            maxLength: 6
      responses:
        '302':
          description: Redirecionamento bem-sucedido
        '404':
          description: URL encurtada não encontrada

  /urls:
    get:
      summary: Lista as URLs encurtadas de um usuário
      description: Este endpoint permite que um usuário autenticado veja suas URLs encurtadas e a quantidade de cliques.
      operationId: listUrls
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de URLs encurtadas do usuário
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    url_original:
                      type: string
                      format: uri
                    url_encurtada:
                      type: string
                    cliques:
                      type: integer
                    criado_em:
                      type: string
                      format: date-time
                    atualizado_em:
                      type: string
                      format: date-time
        '401':
          description: Usuário não autenticado
        '500':
          description: Erro ao listar URLs

  /urls/{id}:
    put:
      summary: Atualiza uma URL encurtada
      description: Este endpoint permite que um usuário autenticado altere a URL associada a um link encurtado.
      operationId: updateUrl
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: O ID da URL a ser atualizada.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                  description: A nova URL para a qual o link encurtado deverá apontar.
      responses:
        '200':
          description: URL atualizada com sucesso
        '400':
          description: Dados inválidos para atualização
        '404':
          description: URL não encontrada
        '500':
          description: Erro ao atualizar a URL

    delete:
      summary: Exclui uma URL encurtada
      description: Este endpoint permite que um usuário autenticado exclua logicamente uma URL encurtada.
      operationId: deleteUrl
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: O ID da URL a ser excluída.
          schema:
            type: integer
      responses:
        '200':
          description: URL excluída com sucesso
        '404':
          description: URL não encontrada
        '500':
          description: Erro ao excluir a URL

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT